refactor(workspace): say what the code does, not what review found
Cleanup of what the diff SAYS. No behaviour change; 464 tests pass and the
ten-mutant catalogue still kills all ten.

Comments state the invariant they protect, in the present tense. Roughly
forty narrated their own history instead — which round found them, what an
earlier version did, how many revisions a claim survived. A comment is read
by someone deciding whether they may change the line below it, and "this
used to be wrong" does not tell them what must stay true. The history is in
the pull request, where it belongs.

Tests are named for the property they assert. Forty-four carried item codes
from the review thread and thirty-two carried ON-LIFT markers explaining how
they had been adapted; the reason for an adaptation belongs in a commit
message, not in the file forever.

The ten test files copied from reviewers are folded into files named for the
question they answer — mutation-guards, undo-and-teardown, config-on-disk,
unbound-and-silence, seam-contract — with each former file's cases and
harness kept in their own describe. Three of them still opened with the
reviewer's own "NOT for commit" header, which is fair evidence they were
copies rather than adaptations. Every case survives; nothing was deduped,
because the mutant catalogue is what proves coverage and dropping cases
without re-running it would be guessing.

Folding surfaced one hollow test of my own. "The snapshot, the write and the
undo all use one resolved path" could not fail: `projectEntry`'s seam never
received the path argument, so dropping it was invisible — the same defect
the restore's seam had. The seam takes it now and the mutant dies.

Also: the version floor loses its shipping-sequencing note, since that
release is out, and a test title loses an issue key that should not be in a
public repo.
